+++ src/inversion-of-control/polaris-types.ts
@@ -1,18 +1,15 @@
 import {
-    InjectableResolver,
-    InjectableType,
     LoggerConfig,
     MiddlewaresConfig,
     PolarisServerConfig,
 } from '../common/injectable-interfaces';
-import { PolarisGraphqlLogger } from '../logging/polaris-graphql-logger';
-import { SchemaCreator } from '../schema/utils/schema-creator';
+import { PolarisGraphQLLogger } from '../logging/polaris-graphql-logger';
 import { GraphQLServer } from '../server/graphql-server';
 
 export const POLARIS_TYPES = {
     GraphQLServer: Symbol('GraphQLServer'),
-    SchemaCreator: Symbol('SchemaCreator'),
-    GraphqlLogger: Symbol('PolarisGraphqlLogger'),
+    GraphQLSchema: Symbol('GraphQLSchema'),
+    GraphQLLogger: Symbol('PolarisGraphQLLogger'),
     PolarisHeadersFactory: Symbol('PolarisHeadersFactory'),
     Middleware: Symbol('Middleware'),
     PolarisServerConfig: Symbol('PolarisServerConfig'),