+++ src/server/graphql-server.ts
@@ -1,21 +1,20 @@
-import { ApolloError, ApolloServer, Config, makeExecutableSchema, PubSub } from 'apollo-server-koa';
+import { ApolloServer, Config, PubSub } from 'apollo-server-koa';
+import { GraphQLError, GraphQLFormattedError, GraphQLSchema } from 'graphql';
 import { applyMiddleware } from 'graphql-middleware';
-import { IResolvers, ITypeDefinitions } from 'graphql-tools';
+import * as http from 'http';
 import { inject, injectable, multiInject } from 'inversify';
 import * as Koa from 'koa';
 import * as koaBody from 'koa-bodyparser';
 import { PolarisServerConfig } from '../common/injectable-interfaces';
+import { IrrelevantEntitiesContainer } from '../common/irrelevant-entities-container';
 import { getHeaders } from '../http/request/polaris-request-headers';
 import { POLARIS_TYPES } from '../inversion-of-control/polaris-types';
-import { PolarisGraphqlLogger } from '../logging/polaris-graphql-logger';
+import { PolarisGraphQLLogger } from '../logging/polaris-graphql-logger';
 import { Middleware } from '../middlewares/middleware';
 import { createMiddleware } from '../middlewares/polaris-middleware-creator';
 import { PolarisProperties } from '../properties/polaris-properties';
-import { SchemaCreator } from '../schema/utils/schema-creator';
+import { IrrelevantEntitiesExtension } from './irrelevant-entities-extension';
 import { PolarisContext } from './polaris-context';
-
-const app = new Koa();
-app.use(koaBody());
 
 export interface GraphQLServer {
     start(): void;
@@ -25,25 +24,20 @@ export type contextCreator = (ctx: Koa.Context) => object;
 
 @injectable()
 export class PolarisGraphQLServer implements GraphQLServer {
-    @inject(POLARIS_TYPES.GraphqlLogger) polarisLogger!: PolarisGraphqlLogger;
+    private app: Koa;
     private server: ApolloServer;
     private polarisProperties: PolarisProperties;
     private customContexts: contextCreator[] = [];
+    private httpServer?: http.Server;
 
     constructor(
-        @inject(POLARIS_TYPES.SchemaCreator) creator: SchemaCreator,
+        @inject(POLARIS_TYPES.GraphQLSchema) schema: GraphQLSchema,
         @inject(POLARIS_TYPES.PolarisServerConfig) propertiesConfig: PolarisServerConfig,
         @multiInject(POLARIS_TYPES.Middleware) middlewares: Middleware[],
+        @inject(POLARIS_TYPES.GraphQLLogger) private polarisLogger: PolarisGraphQLLogger,
     ) {
-        const schema = creator.generateSchema();
-        const executableSchemaDefinition: { typeDefs: ITypeDefinitions; resolvers: IResolvers } = {
-            typeDefs: schema.def,
-            resolvers: schema.resolvers,
-        };
-        const executableSchema = makeExecutableSchema(executableSchemaDefinition);
-
         const executableSchemaWithMiddleware = applyMiddleware(
-            executableSchema,
+            schema,
             ...(middlewares.map(createMiddleware) as any),
         );
         this.polarisProperties = propertiesConfig.polarisProperties;
@@ -53,18 +47,21 @@ export class PolarisGraphQLServer implements GraphQLServer {
             context: (args: { ctx: Koa.Context; connection: any }) => this.getContext(args),
             formatError: (error: any) => this.formatError(error),
             formatResponse: (response: any) => this.formatResponse(response),
+            extensions: [() => new IrrelevantEntitiesExtension()],
         };
         this.server = new ApolloServer(config);
+        this.app = new Koa();
+        this.app.use(koaBody());
+    }
+
+    start(app = this.app) {
         if (this.polarisProperties.endpoint) {
             this.server.applyMiddleware({ app, path: this.polarisProperties.endpoint });
         } else {
             this.server.applyMiddleware({ app });
         }
-    }
-
-    start() {
         const port = this.polarisProperties.port;
-        const httpServer = app.listen(port, () => {
+        this.httpServer = app.listen(port, () => {
             this.polarisLogger.info(
                 `🚀 Server ready at http://localhost:${port}${this.server.graphqlPath}`,
             );
@@ -73,7 +70,13 @@ export class PolarisGraphQLServer implements GraphQLServer {
             );
         });
         if (this.polarisProperties.includeSubscription) {
-            this.server.installSubscriptionHandlers(httpServer);
+            this.server.installSubscriptionHandlers(this.httpServer);
+        }
+    }
+
+    stop() {
+        if (this.httpServer) {
+            this.httpServer.close();
         }
     }
 
@@ -88,16 +91,17 @@ export class PolarisGraphQLServer implements GraphQLServer {
         return response;
     }
 
-    private formatError(error: any) {
+    private formatError(error: GraphQLError): GraphQLFormattedError {
         this.polarisLogger.error('Apollo server error', {
             polarisLogProperties: { throwable: error },
         });
 
-        if (error instanceof ApolloError || error.name === 'GraphQLError') {
-            return { message: error.message, code: error.extensions.code };
-        } else {
-            return new Error('Internal server error');
-        }
+        return {
+            message: error.message,
+            path: error.path,
+            locations: error.locations,
+            extensions: error.extensions,
+        };
     }
 
     private getContext({ ctx, connection }: { ctx: Koa.Context; connection: any }): PolarisContext {
@@ -120,6 +124,7 @@ export class PolarisGraphQLServer implements GraphQLServer {
         const context: PolarisContext = {
             headers,
             body: ctx.request.body,
+            irrelevantEntities: new IrrelevantEntitiesContainer(),
             ...this.getCustomContext(ctx),
         };
         if (this.polarisProperties.includeSubscription) {