+++ test/integration-tests/test-server/config/log-config.ts
@@ -1,15 +1,18 @@
 import { LoggerConfiguration } from '@enigmatis/polaris-logs';
 import { injectable } from 'inversify';
 import { LoggerConfig } from '../../../../src/main';
+import * as polarisLogConfigurationPath from './log-configuration.json';
+
 @injectable()
 export class TestLogConfig implements LoggerConfig {
     loggerConfiguration: LoggerConfiguration;
 
     constructor() {
         this.loggerConfiguration = {
-            loggerLevel: 'debug',
-            logstashConfigurations: [{ logstashHost: '127.0.0.1', logstashPort: 1234 }],
-            writeToConsole: true,
+            loggerLevel: polarisLogConfigurationPath.loggerLevel,
+            logstashConfigurations: polarisLogConfigurationPath.logstashConfigurations,
+            writeToConsole: polarisLogConfigurationPath.writeToConsole,
+            writeFullMessageToConsole: polarisLogConfigurationPath.writeFullMessageToConsole,
         };
     }
 }