+++ test/integration-tests/test-server/schema/resolvers/book-resolvers.ts
@@ -1,32 +1,30 @@
-import { provide } from 'inversify-binding-decorators';
-import { InjectableResolver, POLARIS_TYPES } from '../../../../../src/main';
 import {
     bookByIdQueryResolver,
     bookQueryResolver,
+    bookStartsWithQueryResolver,
     createBookResolver,
     deleteBookResolver,
     deletedBookByIdQueryResolver,
     subscribeResolver,
+    titleResolver,
     updateBookResolver,
 } from './book-resolvers-functions';
 
-@provide(POLARIS_TYPES.InjectableResolver)
-export class BookResolvers implements InjectableResolver {
-    resolver = () => ({
-        Query: {
-            books: bookQueryResolver,
-            bookById: bookByIdQueryResolver,
-            deletedBookById: deletedBookByIdQueryResolver,
+export const resolvers = {
+    Query: {
+        books: bookQueryResolver,
+        booksStartsWith: bookStartsWithQueryResolver,
+        bookById: bookByIdQueryResolver,
+        deletedBookById: deletedBookByIdQueryResolver,
+    },
+    Mutation: {
+        createBook: createBookResolver,
+        updateBook: updateBookResolver,
+        deleteBook: deleteBookResolver,
+    },
+    Subscription: {
+        bookChanged: {
+            subscribe: subscribeResolver,
         },
-        Mutation: {
-            createBook: createBookResolver,
-            updateBook: updateBookResolver,
-            deleteBook: deleteBookResolver,
-        },
-        Subscription: {
-            bookChanged: {
-                subscribe: subscribeResolver,
-            },
-        },
-    });
-}
+    },
+};