+++ test/integration-tests/test-server/schema/types/output/book.ts
@@ -1,0 +1,10 @@
+export const Book = `type Book implements CommonEntity {
+                        id: ID!
+                        creationDate: String,
+                        lastUpdateDate: String,
+                        dataVersion: Int!,
+                        title: String,
+                        author: String,
+                        otherBook: ID,
+                        realityId: Int!
+                     }`;