+++ test/integration-tests/test-server/schema/types/subscription/root-subscription.ts
@@ -1,0 +1,3 @@
+export const Subscription = `type Subscription {
+                                bookChanged(realityId:Int!): Book
+                             }`;