+++ test/middlewares/middleware-activation-condition/filter-data-version.test.ts
@@ -1,4 +1,5 @@
 import { GraphQLResolveInfo } from 'graphql';
+import { IrrelevantEntitiesContainer } from '../../../src/common/irrelevant-entities-container';
 import { ResponseMiddlewareParams } from '../../../src/middlewares/middleware';
 import { DataVersionFilter } from '../../../src/middlewares/middleware-activation-condition/filter-data-version';
 import { PolarisContext } from '../../../src/server/polaris-context';
@@ -9,7 +10,11 @@ const args = {};
 describe('data version tests', () => {
     test('data version header doesnt exist', () => {
         const result = { dataVersion: 0 };
-        const context: PolarisContext = { headers: {}, body: {} };
+        const context: PolarisContext = {
+            headers: {},
+            body: {},
+            irrelevantEntities: new IrrelevantEntitiesContainer(),
+        };
         const middlewareParams: ResponseMiddlewareParams = {
             root: undefined,
             args,
@@ -21,7 +26,11 @@ describe('data version tests', () => {
     });
     describe('data version header exist', () => {
         test('entity data version is bigger than headers data version', () => {
-            const context: PolarisContext = { headers: { dataVersion: 1 }, body: {} };
+            const context: PolarisContext = {
+                headers: { dataVersion: 1 },
+                body: {},
+                irrelevantEntities: new IrrelevantEntitiesContainer(),
+            };
             const middlewareParams: ResponseMiddlewareParams = {
                 root: undefined,
                 args,
@@ -32,7 +41,11 @@ describe('data version tests', () => {
             expect(DataVersionFilter.shouldBeReturned(middlewareParams, false)).toBe(true);
         });
         test('entity data version is equal to headers data version', () => {
-            const context: PolarisContext = { headers: { dataVersion: 1 }, body: {} };
+            const context: PolarisContext = {
+                headers: { dataVersion: 1 },
+                body: {},
+                irrelevantEntities: new IrrelevantEntitiesContainer(),
+            };
             const middlewareParams: ResponseMiddlewareParams = {
                 root: undefined,
                 args,
@@ -43,7 +56,11 @@ describe('data version tests', () => {
             expect(DataVersionFilter.shouldBeReturned(middlewareParams, false)).toBe(false);
         });
         test('entity data version is smaller than headers data version', () => {
-            const context: PolarisContext = { headers: { dataVersion: 2 }, body: {} };
+            const context: PolarisContext = {
+                headers: { dataVersion: 2 },
+                body: {},
+                irrelevantEntities: new IrrelevantEntitiesContainer(),
+            };
             const middlewareParams: ResponseMiddlewareParams = {
                 root: undefined,
                 args,
@@ -54,7 +71,11 @@ describe('data version tests', () => {
             expect(DataVersionFilter.shouldBeReturned(middlewareParams, false)).toBe(false);
         });
         test('entity is sub entity', () => {
-            const context: PolarisContext = { headers: { dataVersion: 2 }, body: {} };
+            const context: PolarisContext = {
+                headers: { dataVersion: 2 },
+                body: {},
+                irrelevantEntities: new IrrelevantEntitiesContainer(),
+            };
             const middlewareParams: ResponseMiddlewareParams = {
                 root: undefined,
                 args,